Fix Dashboard errors
