+++ src/components/App.js
@@ -5,11 +5,10 @@ import { connect } from 'react-redux';
 import LogIn from './auth/LogIn';
 import Dashboard from './Dashboard';
 import FavoriteList from './FavoriteList';
-// import NavBar from './NavBar/NavBar';
 import Home from './Home';
 import AnimeDetails from './AnimeDetails';
 import { isLoggedIn } from '../actions';
-import SideBar from './SideBar';
+import SideBar from './SideBar/SideBar';
 
 const App = ({ isLoggedIn }) => {
   useEffect(() => {
@@ -17,7 +16,6 @@ const App = ({ isLoggedIn }) => {
   }, []);
   return (
     <Router>
-      {/* <NavBar /> */}
       <SideBar />
       <Switch>
         <Route path="/" component={Home} exact />
