--- src/components/NavBar/NavBar.style.css
@@ -1,3 +1,0 @@
-.navBar-container {
-  z-index: 2;
-}
