+++ src/components/SideBar/SideBar.style.css
@@ -10,26 +10,32 @@
   background-color: var(--bg-color);
   box-shadow: rgba(129, 129, 129, 0.356) 10px 5px 10px;
 }
-.side-menu >  li {
+
+.side-menu > li {
   margin-bottom: 1em;
 }
+
 .side-menu a {
   text-decoration: none;
   color: black;
 }
+
 .side-menu button {
   border: none;
 }
+
 .user-container {
   justify-content: center;
   flex-direction: column;
 }
+
 .separator {
   margin-top: auto;
   height: 1px;
   width: 100%;
   background-color: #e9e9e9;
 }
+
 .bg-black {
   background-color: #00000050;
   position: fixed;
@@ -37,6 +43,7 @@
   height: 100vh;
   width: 100vw;
 }
+
 .d-none {
   display: none;
 }
